backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-socketio==5.10.0
scikit-learn==1.3.2
pandas==2.1.3
pyyaml==6.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
